pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
  "hatch-vcs>=0.4",
  "hatchling>=1.25",
]

[project]
name = "hope-documents"
description = ""
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Environment :: Web Environment",
  "Framework :: Django",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.2",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dynamic = [
  "version",
]
dependencies = [
  "click>=8.3",
  "django-admin-extra-buttons>=2.1.1",
  "django-regex>=0.5",
  "django-streaming>=0.8",
  "matplotlib>=3.10.7",
  "opencv-python>=4.12.0.88",
  "pillow>=11.3",
  "pytesseract>=0.3.13",
  "regex>=2025.9.18",
]

scripts.extract = "hope_documents.ocr.__cli__:cli"

[dependency-groups]
dev = [
  "django-browser-reload>=1.19",
  "django-webtest>=1.9.13",
  "factory-boy>=3.3.3",
  "freezegun>=1.5.5",
  "pdbpp>=0.11.7",
  "pre-commit>=4.3",
  "pyquery>=2.0.1",
  "pytest>=8.4.2",
  "pytest-cov>=7",
  "pytest-coverage>=0",
  "pytest-django>=4.11.1",
  "pytest-echo>=2.0.1",
  "pytest-factoryboy>=2.8.1",
  "pytest-xdist>=3.8",
  "responses>=0.25.8",
  "ruff==0.13.1",                # must be aligned with .pre-commit-config.yaml
  "tox>=4.30.2",
]
docs = [
  "markupsafe>=2.1.5",
  "mdx-gh-links>=0.4",
  "mkdocs>=1.6.1",
  "mkdocs-autolinks-plugin>=0.7.1",
  "mkdocs-awesome-pages-plugin>=2.10.1",
  "mkdocs-click>=0.8.1",
  "mkdocs-embed-external-markdown>=3.0.2",
  "mkdocs-ezglossary-plugin>=1.6.10",
  "mkdocs-ezlinks-plugin>=0.1.14",
  "mkdocs-gen-files>=0.5",
  "mkdocs-get-deps>=0.2",
  "mkdocs-include-markdown-plugin>=6.2.2",
  "mkdocs-link-marker>=0.1.3",
  "mkdocs-macros-plugin>=1.0.5",
  "mkdocs-material>=9.5.15",
  "mkdocs-minify-plugin>=0.8",
  "mkdocs-redirects>=1.2.1",
  "mkdocs-simple-hooks>=0.1.5",
  "mkdocstrings[python]>=0.27",
  "pymdown-extensions>=10.14",
]
type-check = [
  "django-stubs[compatible-mypy]",
  "mypy",
  "types-colorama",
  "types-markdown",
  "types-pika",
  "types-requests",
]

[tool.hatch]
build.hooks.vcs.version-file = "src/hope_documents/version.py"
version.source = "vcs"
raw-options = { root = "../.." }

[tool.hatch.build.targets.sdist]
ignore-vcs = true
include = [
  "src/hope_documents/**/*.*",
]

[tool.hatch.build.targets.wheel]
packages = [ "src/hope_documents" ]
